Ask to confirm before rebooting

+++ upgrades/from_any_to_5.3.5/upgrade.sh
@@ -45,6 +45,10 @@ echo -e "${GREEN}Upgrade to v$NEW_PNETLAB_VERSION has been done successfully${NO
 rm -rf upgrade
 #rm $NEW_PNETLAB_VERSION.zip
 
-echo -e "${GREEN}PNETLab VM will be rebooted right now...${NO_COLOR}"
-echo -e "${GREEN}Try to connect again in about a minute${NO_COLOR}"
-reboot
+# confirm reboot
+echo -e "${GREEN}Do you want to reboot now? [y/n]${NO_COLOR}"
+read -r answer
+if [[ $answer == "y" ]]; then
+    echo -e "${GREEN}Try to connect again in about a minute${NO_COLOR}"
+    reboot
+fi
